Refactor Dashboard Page.
