+++ src/components/common/Card.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Card = ({ children, className = '' }) => {
+  return (
+    <div className={`bg-white overflow-hidden shadow rounded-lg ${className}`}>
+      {children}
+    </div>
+  );
+};
+
+export default Card;
