+++ src/pages/Dashboard/components/StatCard.jsx
@@ -1,0 +1,34 @@
+import React from 'react';
+import Card from '../../../components/common/Card';
+
+const StatCard = ({ title, value, icon: Icon, change, trend }) => {
+  return (
+    <Card>
+      <div className="p-5">
+        <div className="flex items-center">
+          <div className="flex-shrink-0 bg-indigo-100 rounded-md p-3">
+            <Icon className="h-6 w-6 text-indigo-600" />
+          </div>
+          <div className="ml-5 w-0 flex-1">
+            <dl>
+              <dt className="text-sm font-medium text-gray-500 truncate">{title}</dt>
+              <dd>
+                <div className="text-lg font-medium text-gray-900">{value}</div>
+              </dd>
+            </dl>
+          </div>
+        </div>
+      </div>
+      <div className="bg-gray-50 px-5 py-3">
+        <div className="text-sm">
+          <span className={`font-medium ${trend === 'up' ? 'text-green-600' : 'text-red-600'}`}>
+            {trend === 'up' ? '↑' : '↓'} {change}
+          </span>{' '}
+          <span className="text-gray-500">from last week</span>
+        </div>
+      </div>
+    </Card>
+  );
+};
+
+export default StatCard;
