+++ src/pages/Dashboard/components/index.js
@@ -1,0 +1,3 @@
+export { default as StatCard } from './StatCard';
+export { default as ChartCard } from './ChartCard';
+export { default as PhysiqueProgressChart } from './PhysiqueProgressChart';
