+++ src/pages/Dashboard/constants.js
@@ -1,0 +1,12 @@
+import { Dumbbell, Apple, Activity, TrendingUp, Scale, Award, Percent, BarChart2 } from 'lucide-react';
+
+export const iconMap = {
+  Scale: Scale,
+  Percent: Percent,
+  Apple: Apple,
+  Activity: Activity,
+  Dumbbell: Dumbbell,
+  Award: Award,
+  BarChart2: BarChart2,
+  TrendingUp: TrendingUp
+};
