+++ src/pages/Dashboard/sections/index.js
@@ -1,0 +1,4 @@
+export { default as StatsSection } from './StatsSection';
+export { default as ChartsSection } from './ChartsSection';
+export { default as PhysiqueProgressSection } from './PhysiqueProgressSection';
+export { default as WorkoutsNutritionSection } from './WorkoutsNutritionSection';
